+++ src/envconfig/index.js
@@ -1,6 +1,6 @@
 let baseURL,imgURL;
 if(process.env.NODE_ENV === 'development'){
-    baseURL = 'http://10.16.201.132:5000/customer/139500'
+    baseURL = 'http://10.16.201.132:5000'
 }else{
     baseURL = 'http://'
 }