+++ src/pages/infomation/info/general/Production.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+export default class Production extends React.Component {
+
+
+    render() {
+        return (
+            <div>
+
+            </div>
+
+        )
+    }
+}