+++ src/router/index.js
@@ -14,6 +14,18 @@ import History from '../pages/infomation/info/exclusive/History';
 import Industry from '../pages/infomation/info/exclusive/Industry';
 import Sentiment from '../pages/infomation/info/exclusive/Sentiment';
 import Relation from '../pages/infomation/info/exclusive/Relation';
+import Strategy from '@/pages/infomation/info/general/Strategy';
+import ShareOrg from '@/pages/infomation/info/general/ShareOrg';
+import Production from '@/pages/infomation/info/general/Production';
+import Dynamic from '@/pages/infomation/info/general/Dynamic';
+import ManagementSearch from '@/pages/infomation/info/general/ManagementSearch';
+
+
+
+
+
+
+
 
 const router = [
     {
@@ -57,7 +69,14 @@ const router = [
             },
             {
                 path: '/info/general',
-                component: General
+                component: General,
+                routes: [
+                    { path: '/info/general/strategy', component: Strategy },
+                    { path: '/info/general/shareOrg', component: ShareOrg },
+                    { path: '/info/general/production', component: Production },
+                    { path: '/info/general/dynamic', component: Dynamic },
+                    { path: '/info/general/managementSearch', component: ManagementSearch },
+                ]
             },
 
         ]
@@ -68,19 +87,19 @@ const router = [
         // exact: true,
         routes: [
             {
-                path:'/collection/collect',
+                path: '/collection/collect',
                 component: Collection,
-                exact:true
+                exact: true
             },
             {
-                path:'/collection/feedback',
+                path: '/collection/feedback',
                 component: Feedback,
-                exact:true
+                exact: true
             },
             {
-                path:'/collection/share',
+                path: '/collection/share',
                 component: Share,
-                exact:true
+                exact: true
             }
         ]