+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.1.0 (2019-04-03)
+
+
+### Features
+
+* **all files:** add first framework ([a07fb7a](https://github.com/ZoneLabyrinth/customer-data/commit/a07fb7a))
+* **create a router component and add to Exclusive Information:** create a router component add add ([443cfec](https://github.com/ZoneLabyrinth/customer-data/commit/443cfec))
+* **exclusive:** completed exclusive page ([6c047a6](https://github.com/ZoneLabyrinth/customer-data/commit/6c047a6))
+* **src:** add exclusive-> basic,history,industry,management,marks tabs ([34f1e31](https://github.com/ZoneLabyrinth/customer-data/commit/34f1e31))
+
+
+
 # 0.1.0 (2019-04-02)